add module and consume breaking change
At some point before they added the go module, go-cdb renamed to cdb.
This still works for GOPATH consumers it is basically a one package
module, and github has left the 302 intact for now, but anybody
migrating to modules breaks. As such, when migrating this repo to
modules, we needed to fix that up.